fix: answer out-of-range catalogue parameters with 400, not 500
Found in production: page=-1, size=0, size=-5 and minPages=-10 returned 500
and logged a stack trace. The controller declares @Min/@Max bounds, but only
conversion failures were mapped; a bound violation raised
ConstraintViolationException and fell through to the catch-all. A very large
page also passed @Min(0) and overflowed page * size into a negative OFFSET.

Constraint violations now map to the same 400 ProblemDetail as other bad
parameters, naming the parameter and nothing else. page is bounded at 10,000
and the offset is computed as a long. Tests cover each case and fail without
the fix.

+++ backend/src/main/java/dev/jorisvrr/reading/catalog/BookSearchQuery.java
@@ -55,7 +55,8 @@ public record BookSearchQuery(
         return q != null && q.length() < SHORT_QUERY_LENGTH;
     }
 
-    public int offset() {
-        return page * size;
+    /** Long, so no page number can wrap the product into a negative OFFSET. */
+    public long offset() {
+        return (long) page * size;
     }
 }
